+++ examples/ui-playground/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @example/ui-playground
 
+## 0.1.0-alpha.17
+
+### Patch Changes
+
+- Updated dependencies [[`0b9269d`](https://github.com/softnetics/genseki/commit/0b9269dcc3830afc395363d7f1e94cecd33e6b03)]:
+  - @genseki/react@0.1.0-alpha.18
+  - @genseki/next@0.1.0-alpha.18
+  - @genseki/react-query@0.1.0-alpha.18
+  - @genseki/rest@0.1.0-alpha.18
+
+## 0.1.0-alpha.16
+
+### Minor Changes
+
+- [#79](https://github.com/softnetics/genseki/pull/79) [`082b5bc`](https://github.com/softnetics/genseki/commit/082b5bc747adff6c0adc5526f80b8ea2aedd0b8e) Thanks [@Desmenez](https://github.com/Desmenez)! - redesign components and add new component
+
+- [#79](https://github.com/softnetics/genseki/pull/79) [`06da61b`](https://github.com/softnetics/genseki/commit/06da61bd96f18d690eb72ef3bf2fa919c3dae4a6) Thanks [@Desmenez](https://github.com/Desmenez)! - add darkmode switch button
+
+### Patch Changes
+
+- Updated dependencies [[`082b5bc`](https://github.com/softnetics/genseki/commit/082b5bc747adff6c0adc5526f80b8ea2aedd0b8e), [`f2adbb5`](https://github.com/softnetics/genseki/commit/f2adbb590d9a4795eb4afebaad032cf39f25aa8d)]:
+  - @genseki/react@0.1.0-alpha.17
+  - @genseki/next@0.1.0-alpha.17
+  - @genseki/react-query@0.1.0-alpha.17
+  - @genseki/rest@0.1.0-alpha.17
+
+## 0.1.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [[`b56cc5e`](https://github.com/softnetics/genseki/commit/b56cc5e2135a7e15cccb0595fc266d329e027a87), [`63c94ed`](https://github.com/softnetics/genseki/commit/63c94ed9ff12730d2fa6121f6d3215ec906aff23), [`7acbdc9`](https://github.com/softnetics/genseki/commit/7acbdc902aace63590a0a45f8ac9d67a57c1c7de)]:
+  - @genseki/react@0.1.0-alpha.16
+  - @genseki/next@0.1.0-alpha.16
+  - @genseki/react-query@0.1.0-alpha.16
+  - @genseki/rest@0.1.0-alpha.16
+
+## 0.1.0-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [[`9ad485e`](https://github.com/softnetics/genseki/commit/9ad485eedfd3e890d8f58158b7c7b25b27cdbe87)]:
+  - @genseki/react@0.1.0-alpha.15
+  - @genseki/next@0.1.0-alpha.15
+  - @genseki/react-query@0.1.0-alpha.15
+  - @genseki/rest@0.1.0-alpha.15
+
 ## 0.1.0-alpha.13
 
 ### Patch Changes